user-api/requirements.txt
fastapi[all]
pandas
sqlalchemy
psycopg2
python-jose
passlib
bcrypt